Remove ContextError and return context errors directly (#17)
Prior to this commit every context error returned by the library was wrapped in
a custom struct called ContextError. As a result, checking cancellation errors
returned from the library with errors.Is(err, ctx.Err()) would return false.
This design decision was made before the existence of errors.Is(), but is no
longer helpful. With this commit, any context errors are returned directly
without embedding in a custom struct.